refactor(transformService): separate service into separate files per responsibility
