+++ server/services/transform-service/index.js
@@ -1,0 +1,13 @@
+'use strict';
+
+const _ = require('lodash');
+
+const { transformResponse } = require('./response');
+
+module.exports = () => ({
+	response(settings, ctx) {
+		if (_.has(settings, ['responseTransforms'])) {
+			transformResponse(settings.responseTransforms, ctx);
+		}
+	},
+});
