+++ server/services/transform-service/util.js
@@ -1,0 +1,10 @@
+function removeObjectKey(object, key) {
+	return {
+		id: object.id,
+		...object[key],
+	};
+}
+
+module.exports = {
+	removeObjectKey,
+};
